% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cv.R
\name{cvLocal_PAV_dtr}
\alias{cvLocal_PAV_dtr}
\title{Function to compute the Local Population Average Value under SMART for cross-validated DTR.}
\usage{
cvLocal_PAV_dtr(Acv, Dhatcv, Ycv, indcv, s, centered = FALSE)
}
\arguments{
\item{Acv}{A matrix of the actual treatment. Each row represents a unit-level randomized sequential assignments.}

\item{Dhatcv}{A list of DTR matrix derived from each trail of cross-validation.}

\item{Ycv}{A vector of the final outcome variable of interest for each sample.}

\item{indcv}{A vector of integers (between 1 and number of folds inclusive) indicating which testing set does each sample belong to.}

\item{s}{The stage up to which DTR should be evaluated..}

\item{centered}{If \code{TRUE}, the outcome variables would be centered before processing. This minimizes
the variance of the estimator. Default is \code{FALSE}.}
}
\value{
A list that contains the following items: \item{stage}{The stage up to which DTR has been evaluated}，
\item{lpav}{The estimated local population value of the estimated DTR.},
\item{sd}{The standard deviation of the estimated local population value.}
}
\description{
Given a A,D,Y dataset, this function estimates the Population Average Value of an DTR derived from the same data by CV up to stage s.
}
\examples{
Acv <- matrix(rep(0,1000*2),nrow = 1000,ncol= 2)
for(i in 1:2){
 n1=round(1000/2)
 n0=1000-n1
 ind=sample(1:1000,size=n1)
 Acv[ind,i] = 1}
Dhatcv <- list()
for (i in 1:5) {
D <- matrix(0, nrow = 1000, ncol = 2)
num_ones <- 500 + (i - 1)
D[1:num_ones, 2] <- 1
Dhatcv [[i]] <- D
}
Ycv =  rnorm(1000,3,1.5)
indcv <- sample(rep(1:5, each = 200))
lpav_est <- cvLocal_PAV_dtr(Acv,Dhatcv,Ycv,indcv,2,centered = FALSE)
lpav_est$lpav
lpav_est$sd
}
