% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cv.R
\name{cvPAPE_dtr}
\alias{cvPAPE_dtr}
\title{Function to compute the Population Average Prescriptive Effect under SMART for cross-validated DTR.}
\usage{
cvPAPE_dtr(Acv, Dhacvt, Ycv, indcv, centered = FALSE)
}
\arguments{
\item{Acv}{A matrix of the actual treatment. Each row represents a unit-level randomized sequential assignments.}

\item{Ycv}{A vector of the final outcome variable of interest for each sample.}

\item{indcv}{A vector of integers (between 1 and number of folds inclusive) indicating which testing set does each sample belong to.}

\item{centered}{If \code{TRUE}, the outcome variables would be centered before processing. This minimizes
the variance of the estimator. Default is \code{FALSE}.}

\item{Dhatcv}{A list of DTR matrix derived from each trail of cross-validation.}
}
\value{
A list that contains the following items: \item{pape}{The estimated pape of the estimated DTR.},
\item{sd}{The standard deviation of the estimated pape.}
}
\description{
Given a A,D,Y dataset, this function estimates the  Population Average Prescriptive Effect of an DTR derived from the same data by CV.
}
\examples{
Acv <- matrix(rep(0,1000*2),nrow = 1000,ncol= 2)
for(i in 1:2){
 n1=round(1000/2)
 n0=1000-n1
 ind=sample(1:1000,size=n1)
 Acv[ind,i] = 1}
Dhatcv <- list()
for (i in 1:5) {
D <- matrix(0, nrow = 1000, ncol = 2)
num_ones <- 500 + (i - 1)
D[1:num_ones, 2] <- 1
Dhatcv [[i]] <- D
}
Ycv =  rnorm(1000,3,1.5)
indcv <- sample(rep(1:5, each = 200))
pape_est <- cvPAPE_dtr(Acv,Dhatcv,Ycv,indcv,centered = FALSE)
pape_est$pape
pape_est$sd
}
