% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunction.R
\name{dtrMatrix}
\alias{dtrMatrix}
\title{indicator of DTR}
\usage{
dtrMatrix(D_i, s)
}
\arguments{
\item{D_i}{A vector of DTR.}

\item{s}{time to start randomization}
}
\value{
DTR vector up to time s
}
\description{
indicator of DTR
}
