% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fixed.R
\name{Local_PAPE_dtr}
\alias{Local_PAPE_dtr}
\title{Function to compute the Local Population Average Prescriptive Effect under SMART.}
\usage{
Local_PAPE_dtr(A, D, Y, s)
}
\arguments{
\item{A}{A matrix of the actual treatment. Each row represents a unit-level randomized sequential assignments.}

\item{D}{A matrix of the unit-level dynamic treatment that would have been assigned by the
individualized dynamic treatment rule.}

\item{Y}{A vector of the final outcome variable of interest for each sample.}

\item{s}{The stage up to which DTR should be evaluated..}
}
\value{
A list that contains the following items: \item{stage}{The stage up to which DTR has been evaluated}，
\item{lpape}{The estimated local population average prescriptive effect of given DTR.},
\item{sd}{The standard deviation of the estimated local population average prescriptive effect .}，
\item{cov}}{covariance term in the variance formula (for internal use).},\item{St}{E\link{S_{t}^2}/n_t (for internal use).}
}
\description{
Given a A,D,Y dataset and stage s, this function estimates the Population Average Prescriptive Effect of a given DTR up stage s.
}
\examples{
A <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
for(i in 1:2){
 n1=round(100/2)
 n0=100-n1
 ind=sample(1:100,size=n1)
 A[ind,i] = 1}
D <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
D[1:50,2] = 1
Y =  rnorm(100,3,1.5)
lpape_est <- Local_PAPE_dtr(A,D,Y,2)
lpape_est$lpape
lpape_est$sd
}
