% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fixed.R
\name{PAPE_dtr}
\alias{PAPE_dtr}
\title{Function to compute the Population Average Prescriptive Effect under SMART.}
\usage{
PAPE_dtr(A, D, y)
}
\arguments{
\item{A}{A matrix of the actual treatment. Each row represents a unit-level randomized sequential assignments.}

\item{D}{A matrix of the unit-level dynamic treatment that would have been assigned by the
individualized dynamic treatment rule.}

\item{y}{A vector of the final outcome variable of interest for each sample.}
}
\value{
A list that contains the following items:
\item{pape}{The estimated population average prescriptvie effect of a given DTR.},
\item{sd}{The standard deviation of the estimated pape.},
\item{st}{E\link{S_{t}^2}/n_t (for internal use).},
\item{cov}{covariance term in the variance formula (for internal use).},
\item{sate}{vector of sample average value of each possible assignment under actual treatment(for internal use).},
\item{sav}{vector of sample average value of each possible assignment under dtr (for internal use).},
\item{rpav}{vector of ramdomized average value of each possible assignment under actual treatment (for internal use).},
\item{sape}{vector of sample pape of each possible assignment under dtr (for internal use).},
}
\description{
Given a A,D,Y dataset, this function estimates Population Average Prescriptive Effect of a given DTR.
}
\examples{
A <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
for(i in 1:2){
 n1=round(100/2)
 n0=100-n1
 ind=sample(1:100,size=n1)
 A[ind,i] = 1}
D <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
D[1:50,2] = 1
Y =  rnorm(100,3,1.5)
pape_est <- PAPE_dtr(A,D,Y)
pape_est$pape
pape_est$sd
}
