% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fixed.R
\name{PAV_dtr}
\alias{PAV_dtr}
\title{Function to compute the Population Average Value under SMART.}
\usage{
PAV_dtr(A, D, y)
}
\arguments{
\item{A}{A matrix of the actual treatment. Each row represents a unit-level randomized sequential assignments.}

\item{D}{A matrix of the unit-level dynamic treatment that would have been assigned by the
individualized dynamic treatment rule.}

\item{y}{A vector of the final outcome variable of interest for each sample.}
}
\value{
A list that contains the following items: \item{sav}{sample average value for each possible treatment}
\item{pav}{The estimated population value of a DTR.}, \item{sd}{The standard deviation of the estimated population value.}
}
\description{
Given a A,D,Y dataset, this function estimates the Population Average Value of a given DTR.
}
\examples{
A <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
for(i in 1:2){
 n1=round(100/2)
 n0=100-n1
 ind=sample(1:100,size=n1)
 A[ind,i] = 1}
D <- matrix(rep(0,100*2),nrow = 100,ncol= 2)
D[1:50,2] = 1
Y =  rnorm(100,3,1.5)
pav_est <- PAV_dtr(A,D,Y)
pav_est$pav
pav_est$sd
}
