% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunction.R
\name{trtMatrix}
\alias{trtMatrix}
\title{Indicator vector showing if Treatment align with DTR}
\usage{
trtMatrix(A_i, D_i, s = NULL)
}
\arguments{
\item{A_i}{A vector of actual treatment.}

\item{D_i}{A vector of DTR.}

\item{s}{stage to start randomization (when not to consider D)
If s=NULL, if Assinment follow DTR till the end; if s=1 Indicator of actual treatment from the start.
if s = 2, Indicator of follwing DTR up to the first stage and then actual treatment from stage 2.}
}
\value{
A matrix of treatment indicator.
}
\description{
Indicator vector showing if Treatment align with DTR
}
